Update plex_webhook_release.yaml
Added back in album, it turns out plex does not send out album tag when the new music being added is a new artist not already in the library. 

Will need to compare the webhook data received between new artist and album.

But for now Artist and Album are both being recognised.